chore(packages/spacings): move spacings to packages
