+++ packages/components/spacings/inline/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './inline';
+
+export { default as version } from './version';
