+++ packages/components/spacings/inline/src/inline.js
