+++ packages/components/spacings/inline/src/inline.story.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, select } from '@storybook/addon-knobs/react';
 import styled from '@emotion/styled';
-import Text from '../../../../packages/components/text';
-import Inset from '../inset';
+import Text from '../../../text';
+import Inset from '../../inset';
 import Inline from './inline';
-import Readme from './README.md';
+import Readme from '../README.md';
 
 const Stack = styled.div`
   > * + * {
