+++ packages/components/spacings/inset-squish/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './inset-squish';
+
+export { default as version } from './version';
