+++ packages/components/spacings/inset/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './inset';
+
+export { default as version } from './version';
