+++ packages/components/spacings/stack/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './stack';
+
+export { default as version } from './version';
