+++ packages/components/spacings/stack/src/stack.story.js
@@ -3,10 +3,10 @@ import { storiesOf } from '@storybook/react';
 import { withKnobs, select } from '@storybook/addon-knobs/react';
 import styled from '@emotion/styled';
 import Text from '@commercetools-uikit/text';
-import Inline from '../inline';
-import Inset from '../inset';
+import Inline from '../../inline';
+import Inset from '../../inset';
 import Stack from './stack';
-import Readme from './README.md';
+import Readme from '../README.md';
 
 const StackColorWrapper = styled.div`
   background-color: #d4e0ec;
