--- src/components/spacings/inline/index.js
@@ -1,1 +1,0 @@
-export { default } from './inline';
