--- src/components/spacings/inset-squish/index.js
@@ -1,1 +1,0 @@
-export { default } from './inset-squish';
