--- src/components/spacings/inset/index.js
@@ -1,1 +1,0 @@
-export { default } from './inset';
