--- src/components/spacings/stack/index.js
@@ -1,1 +1,0 @@
-export { default } from './stack';
